feat(application): add create question use case
Implements the 'CreateQuestionUseCase' responsible for handling the creation of new questions.

This use case receives the title, content, and authorId, instantiates a new 'Question' entity (which also automatically generates the 'slug' based on the title), and persists it using the 'QuestionsRepositories'.

Unit tests are included using an in-memory repository to validate that the question is created and stored correctly.