feat(application): add comment on question use case
Implements the 'CommentOnQuestionUseCase', responsible for allowing a user to add a comment to a specific question.

This use case validates one key rule:
1. It checks if the target question exists using the 'QuestionsRepository'.

If the question is found, the use case creates a new 'QuestionComment' entity and persists it using the 'QuestionCommentsRepository'.

Unit tests are included using in-memory repositories to validate the successful creation of a comment on an existing question.